did some restructuring, helper methods are now in a separate file so that unit tests can be performed. Also added the ability to limit the time piper plays audio with the play function to an integer number of seconds. If the last argument called with the play function is an integer it will only play the first however many seconds of the audio
