+++ good_bot.txt
@@ -1,1 +1,1 @@
-2
+6
